feat: Add posts for next dynamic routing
