--- src/pages/series/list.js
@@ -1,12 +1,0 @@
-import React from "react";
-import SeriesPostList from "@components/series/SeriesPostList";
-
-function list() {
-  return (
-    <div>
-      <SeriesPostList />
-    </div>
-  );
-}
-
-export default list;
